Fix bug in parsing \E.
There was a bug in parsing the part of a regexp that comes after \E.
For example, "\Qabc\E.*" was parsed to be equivalent to  "abc." (or
"abc.(?:)*" to be exact), which is obviously incorrect.

Thanks to Bill Yerazunis <wsy@merl.com> for reporting the bug.

darcs-hash:20050328185603-ced27-23246308f3c1926f51c111b78207e1536a202452.gz