Fix str_handler_get_next in to inspect 'ch', not 'c'.
This function was always returning 0.  Also, TRE relies on the fact
that this function sets *c to 0, so I'm adding that to the API
documentation.  If you change str_handler_get_next and return 1 if ch
is 0, before setting *c, you will get an infinite loop or a seg fault.
The other alternative is to fix GET_NEXT_WCHAR.  One advantage of
fixing GET_NEXT_WCHAR is that the data can be allowed to contain the
null character and so it can be used for binary data.

darcs-hash:20090318180040-54bb4-93e78d4e2a05c4311757f325aca38ae3e9fa1937.gz