Add a test case for bounded iterations.
This test case exposes a bug which causes the upper limit of a finite
bound to be ignored in certain cases.

The regular expression used in the test case is a minimal reproducer;
a more practical but less readable and less easily debugged example
would be a regular expression that accepts fully-qualified domain
names.