Avoid use-after-free warning
More recent versions of GCC try to check for use after free. Realloc
will take the input pointer, `ptrn` and with free it after a successful
allocation. However, it will not change the contents of the pointer, and
thus GCC in some cases thinks it can/could be used. By explicitly
setting `ptrn` to NULL after an successful allocation, we can keep GCC
happy and our codebase sane.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>