Postpone position assignment until the end.
During AST expansion, it is possible for new literal nodes to be created
and given the same position as existing literals.  The code attempts to
reposition nodes when this happens but does not always succeed.  The
result is a short-circuit in the TNFA which can manifest as an infinite
loop, e.g. a bounded iteration being treated as an unbounded one.  This
reshuffling can also leave positions unfilled, wasting memory.

Instead of trying to assign positions to literal nodes on creation, and
then having to renumber them on the fly during AST expansion, wait until
the end and move position assignment into `tre_compute_nfl()`, which we
rename to `tre_compute_npfl()`.