regexec: Fix unintended return.
In `tre_match()`, there is a spot where we check if the provided source
is capable of rewinding, and return if it isn't.  Due to missing braces,
that return becomes unconditional when `TRE_USE_ALLOCA` is false.  As a
result, calling `regexec()` with a regex that includes back references
always fails with `REG_BADPAT`.

Fixes: 6f09108bb17c