Fix BRE parentheses
Previously, when the parser encountered a left or right parenthesis while
parsing an atom, it would check if it was either in ERE mode or if the
preceding character was a backslash.  This is unreliable as we cannot
know at this point if said backslash was itself escaped.  Instead, when
we encounter a backslash (which we know is not escaped), and are not in
ERE mode, check if the next character is a left or right parenthesis,
and if so, jump directly to the appropriate point.